--- src/components/DiaryListPage/Diary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Diary';
